+++ packages/isolate-protocol/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ricsam/isolate-protocol
 
+## 0.1.16
+
+### Patch Changes
+
+- fix abort fetch
+
 ## 0.1.15
 
 ### Patch Changes